updated parsing for port number; moved from server to config_parser

+++ include/config_var.h
@@ -6,15 +6,15 @@
 
 struct config_var
 {
-  config_var()
-  {
-    port = -1;
-    echo_roots = std::vector<std::string>(1, "");
-    static_roots = std::vector<std::string>(1, "");
-  }
-  short int port;
-  std::vector<std::string> echo_roots;
-  std::vector<std::string> static_roots;
+    config_var()
+    {
+        port = -1;
+        echo_roots = std::vector<std::string>(1,"");
+        static_roots = std::vector<std::string>(1,"");
+    }
+    int port;
+    std::vector<std::string> echo_roots;
+    std::vector<std::string> static_roots;
 };
 
 #endif
